20210807
1. 补全Permission Controller的接口
2. 修改permission表role字段为unique字段
3. Remove redundant unchecked supression warnings in GlobalExceptionHandler.class
4. 修复并精简auth模块所有接口 去掉不必要的检查 并重新测试
5. 移除Spring Security
6. 移除RedisFeignService 修复RedisService接口

+++ MangoLee-CMS-auth/src/main/java/org/mangolee/controller/PermissionController.java
@@ -57,23 +57,13 @@ public class PermissionController {
             @NotNull String role
     ) {
         try {
-            QueryWrapper<Permission> wrapper = new QueryWrapper<Permission>()
-                    .eq("role", role);
-            Permission permission = permissionService.getOne(wrapper);
-            // 检查是否存在此权限 如果存在则抛出异常
-            if (permission != null) {
-                throw new BaseException(Result.BAD_REQUEST);
-            }
-            permission = new Permission();
+            Permission permission = new Permission();
             permission.setRole(role);
             if (!permissionService.save(permission)) {
                 throw new BaseException(Result.BAD_REQUEST);
             }
+            QueryWrapper<Permission> wrapper = new QueryWrapper<Permission>().eq("role", role);
             permission = permissionService.getOne(wrapper);
-            // 若创建的权限找不到则抛出异常
-            if (permission == null) {
-                throw new BaseException(Result.BAD_REQUEST);
-            }
             return Result.success(permission);
         } catch (Exception e) {
             throw new BaseException(Result.BAD_REQUEST);
@@ -88,7 +78,7 @@ public class PermissionController {
             @NotNull Long id,
             @ApiParam(value = "权限角色", required = true)
             @PathVariable("role")
-            @NotNull String role
+            @NotNull String newRole
     ) {
         try {
             // 判断权限是否为null
@@ -98,14 +88,16 @@ public class PermissionController {
             }
             String oldRole = permission.getRole();
             // 更新权限名称
-            permission = new Permission();
-            permission.setId(id);
-            permission.setRole(role);
-            if (!permissionService.save(permission)) {
+            permission.setRole(newRole);
+            permission.setGmtModified(null);
+            if (!permissionService.updateById((permission))) {
                 throw new BaseException(Result.BAD_REQUEST);
             }
-            // 更新User表中的对应权限
-            userService.saveBatchByRole(oldRole, role);
+            // 批量更新User
+            User user = new User();
+            user.setRole(newRole);
+            QueryWrapper<User> wrapper = new QueryWrapper<User>().eq("role", oldRole);
+            userService.update(user, wrapper);
             permission = permissionService.getById(id);
             return Result.success(permission);
         } catch (Exception e) {
@@ -115,28 +107,20 @@ public class PermissionController {
 
     @ApiOperation("根据主键ID进行逻辑删除")
     @DeleteMapping("/logicaldeletebyid/{id}")
-    public Result<Permission> logicalDeleteById(
+    public Result<Void> logicalDeleteById(
             @ApiParam(value = "主键ID", required = true)
             @PathVariable("id")
             @NotNull Long id) {
         try {
+            Permission permission = permissionService.getById(id);
+            if (permission == null) {
+                throw new BaseException(Result.BAD_REQUEST);
+            }
+            String role = permission.getRole();
             if (!permissionService.removeById(id)) {
                 throw new BaseException(Result.BAD_REQUEST);
             }
-            return Result.success();
-        } catch (Exception e) {
-            throw new BaseException(Result.BAD_REQUEST);
-        }
-    }
-
-    @ApiOperation("根据主键ID进行物理删除")
-    @DeleteMapping("/physicaldeletebyid/{id}")
-    public Result<Permission> physicalDeleteById(
-            @ApiParam(value = "主键ID", required = true)
-            @PathVariable("id")
-            @NotNull Long id) {
-        try {
-            permissionService.physicalDeleteById(id);
+            userService.updateRoleBatchWithNull(role);
             return Result.success();
         } catch (Exception e) {
             throw new BaseException(Result.BAD_REQUEST);
@@ -145,30 +129,37 @@ public class PermissionController {
 
     @ApiOperation("根据角色权限进行逻辑删除")
     @DeleteMapping("/logicaldeletebyrole/{role}")
-    public Result<Permission> logicalDeleteByRole(
+    public Result<Void> logicalDeleteByRole(
             @ApiParam(value = "权限角色", required = true)
             @PathVariable("role")
             @NotNull String role) {
         try {
-            QueryWrapper<Permission> wrapper = new QueryWrapper<Permission>()
-                    .eq("role", role);
+            QueryWrapper<Permission> wrapper = new QueryWrapper<Permission>().eq("role", role);
             if (!permissionService.remove(wrapper)) {
                 throw new BaseException(Result.BAD_REQUEST);
             }
+            userService.updateRoleBatchWithNull(role);
             return Result.success();
         } catch (Exception e) {
             throw new BaseException(Result.BAD_REQUEST);
         }
     }
 
-    @ApiOperation("根据角色权限进行物理删除")
-    @DeleteMapping("/physicaldeletebyrole/{role}")
-    public Result<Permission> physicalDeleteByRole(
-            @ApiParam(value = "权限角色", required = true)
-            @PathVariable("role")
-            @NotNull String role) {
+    @ApiOperation("根据主键ID进行物理删除")
+    @DeleteMapping("/physicaldeletebyid/{id}")
+    public Result<Void> physicalDeleteById(
+            @ApiParam(value = "主键ID", required = true)
+            @PathVariable("id")
+            @NotNull Long id) {
         try {
-            permissionService.physicalDeleteByRole(role);
+            Permission permission = permissionService.getPermissionByIdIgnoreLogicalDeletion(id);
+            if (permission == null) {
+                throw new BaseException(Result.BAD_REQUEST);
+            }
+            String role = permission.getRole();
+            permissionService.physicalDeleteById(id);
+            // User表只对未被逻辑删除的条目进行更新
+            userService.updateRoleBatchWithNull(role);
             return Result.success();
         } catch (Exception e) {
             throw new BaseException(Result.BAD_REQUEST);
